+++ PromptBIMTestApp1/PromptBIMTestApp1App.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+@main
+struct PromptBIMTestApp1App: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+        .windowStyle(.titleBar)
+        .defaultSize(width: 1200, height: 800)
+    }
+}